requirements.txt
# Core
transformers==4.40.0
sentence-transformers==2.2.2
torch>=2.0.0
faiss-cpu==1.7.4             # if you have no GPU; if you have GPU you can swap to faiss-gpu
flask==2.2.5
numpy==1.26.0
scikit-learn==1.3.0
uvicorn==0.23.0             # optional, if you prefer uvicorn for running the API
# NOTE: version pinning is indicative – adjust for your environment.
